// src/app/shared/models/nav-item-model.component.ts
export class NavItemModelComponent{
  href: string;
  routerLink: string;
  fragment: string;
  target: string;
  text: string;
  html: string;
  id: string;
  active: boolean;
  disabled: boolean;
  divider: boolean;
  classes: string;

  role: string;
  ariaLabel: string;
  ariaDescribedBy: string;
  ariaLabelledBy: string;
  ariaHidden: string;
  ariaDisabled: string;
  ariaControls: string;
  ariaCurrent: string;
  ariaLive: string;
  ariaExpanded: string;
  ariaErrorMessage: string;
  ariaHasPopup: string;
}